refac: 웹 접근성을 위한 키보드 호환 코드 분리 및 적용 외
